refactor: Consolidate utility functions and improve component props structure
