+++ dashboard/src/components/NodeHealth/__tests__/utils.ts
@@ -1,0 +1,25 @@
+export const formatBytes = (bytes: number) => {
+  const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
+  if (bytes === 0) return '0 B';
+  const i = Math.floor(Math.log(bytes) / Math.log(1024));
+  return Math.round((bytes / Math.pow(1024, i)) * 100) / 100 + ' ' + sizes[i];
+};
+
+export const formatDuration = (ms: number) => {
+  const seconds = Math.floor(ms / 1000);
+  const minutes = Math.floor(seconds / 60);
+  const hours = Math.floor(minutes / 60);
+  const days = Math.floor(hours / 24);
+
+  if (days > 0) return `${days}d ${hours % 24}h`;
+  if (hours > 0) return `${hours}h ${minutes % 60}m`;
+  if (minutes > 0) return `${minutes}m ${seconds % 60}s`;
+  return `${seconds}s`;
+};
+export const TABS = [
+  { label: 'Blockchain', value: 'blockchain' },
+  { label: 'Peers', value: 'peers' },
+  { label: 'Network', value: 'network' },
+  { label: 'Mempool', value: 'mempool' },
+  { label: 'Bandwidth', value: 'bandwidth' },
+];
